requirements.txt
fastapi==0.115.9
uvicorn==0.34.2
python-dotenv==1.1.0
google-generativeai==0.8.5
supabase==2.15.1
langchain==0.3.25
langchain-community==0.3.24
chromadb==1.0.9
sentence-transformers==4.1.0
requests==2.32.3
transformers==4.51.3
huggingface-hub==0.31.2
protobuf==5.29.4
pydantic==2.11.4
typing-extensions==4.13.2
